fix the termiation of previous run

+++ build/run-e2e-tests.sh
@@ -77,7 +77,7 @@ echo -e "\nApply channel CRDs\n"
 kubectl apply -f deploy/crds
 
 if [ "$TRAVIS_BUILD" != 1 ]; then
-    echo -e "\nwait for pod to be ready\n"
+    echo -e "\nWait for pod to be ready\n"
     sleep 35
 fi
 
@@ -116,7 +116,9 @@ E2E_DATA_PATH="applifecycle-backend-e2e/default-e2e-test-data"
 echo -e "\nTerminate the running test server\n"
 ps aux | grep 8765 | grep -v 'grep'
 
-ps aux | grep ${IMG} | grep -v 'grep' | awk '{print $2}' | xargs kill -9
+if [ "$TRAVIS_BUILD" != 1 ]; then
+    ps aux | grep ${IMG} | grep -v 'grep' | awk '{print $2}' | xargs kill -9
+fi
 
 ${E2E_BINARY_PATH}/applifecycle-backend-e2e -cfg cluster_config -data ${E2E_DATA_PATH} &
 
@@ -124,6 +126,8 @@ sleep 10
 curl http://localhost:8765/cluster
 
 echo -e "\nTerminate the test server\n"
-ps aux | grep ${IMG} | grep -v 'grep' | awk '{print $2}' | xargs kill -9
+if [ "$TRAVIS_BUILD" != 1 ]; then
+    ps aux | grep ${IMG} | grep -v 'grep' | awk '{print $2}' | xargs kill -9
+fi
 
 exit 0;
